refactor(tools): dedupe cloud-sql-connect-gce helpers into cloudsqlconnect
Move getComputeService, extractSQLInfo, extractVMInfo, findVM, and the
package-level compute-service singleton into internal/util/cloudsqlconnect/gce.go
as exported GetComputeService, ExtractSQLInfo, ExtractVMInfo, FindVM. All
three connect-gce tool packages (postgres, mysql, mssql) previously carried
byte-identical copies (~450 lines total); each tool file drops from ~372 to
~230 lines.

Also fixes a reliability bug flagged by review: the compute.NewService
initializer inside sync.Once was using the request-scoped ctx from the first
Invoke — if that first call cancelled or timed out, the cached error would
poison every subsequent invocation. The shared GetComputeService now
initializes with context.Background(); per-call cancellation still flows
through Instances.Get(...).Context(ctx).Do().

Fixes gemini-code-assist review findings on #3740.
